user type edited

+++ project/main.py
@@ -21,39 +21,51 @@ def showRestaurants():
 @main.route('/restaurant/new/', methods=['GET','POST'])
 @login_required
 def newRestaurant():
-  if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'], user_id = current_user.id)
-      db.session.add(newRestaurant)
-      flash('New Restaurant %s Successfully Created' % newRestaurant.name)
-      db.session.commit()
-      return redirect(url_for('main.showRestaurants'))
-  else:
-      return render_template('newRestaurant.html')
+    if current_user.user_type != 'customer':
+        if request.method == 'POST':
+            newRestaurant = Restaurant(name = request.form['name'], user_id = current_user.id)
+            db.session.add(newRestaurant)
+            flash('New Restaurant %s Successfully Created' % newRestaurant.name)
+            db.session.commit()
+            return redirect(url_for('main.showRestaurants'))
+        else:
+            return render_template('newRestaurant.html')
+    else: 
+       return redirect(url_for('main.showRestaurants'))
 
 #Edit a restaurant
 @main.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
-  editedRestaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-      if request.form['name']:
-        editedRestaurant.name = request.form['name']
-        flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
-        return redirect(url_for('main.showRestaurants'))
-  else:
-    return render_template('editRestaurant.html', restaurant = editedRestaurant)
+    restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if current_user.name == 'Admin' or current_user.id == restaurant.user_id:
+        editedRestaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+        if request.method == 'POST':
+            if request.form['name']:
+                editedRestaurant.name = request.form['name']
+                flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
+                return redirect(url_for('main.showRestaurants'))
+        else:
+            return render_template('editRestaurant.html', restaurant = editedRestaurant)
+    else: 
+       return redirect(url_for('main.showRestaurants'))    
+
 
 
 #Delete a restaurant
 @main.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
-  restaurantToDelete = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-    db.session.delete(restaurantToDelete)
-    flash('%s Successfully Deleted' % restaurantToDelete.name)
-    db.session.commit()
-    return redirect(url_for('main.showRestaurants', restaurant_id = restaurant_id))
-  else:
-    return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
+    restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if current_user.name == 'Admin' or current_user.id == restaurant.user_id:
+        restaurantToDelete = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+        if request.method == 'POST':
+            db.session.delete(restaurantToDelete)
+            flash('%s Successfully Deleted' % restaurantToDelete.name)
+            db.session.commit()
+            return redirect(url_for('main.showRestaurants', restaurant_id = restaurant_id))
+        else:
+            return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
+    else: 
+       return redirect(url_for('main.showRestaurants'))   
 
 #Show a restaurant menu
 @main.route('/restaurant/<int:restaurant_id>/')
@@ -70,15 +82,18 @@ def showMenu(restaurant_id):
 #Create a new menu item
 @main.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
-  restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
-      db.session.add(newItem)
-      db.session.commit()
-      flash('New Menu %s Item Successfully Created' % (newItem.name))
-      return redirect(url_for('main.showMenu', restaurant_id = restaurant_id))
-  else:
-      return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+    restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if current_user.name == 'Admin' or current_user.id == restaurant.user_id:
+        if request.method == 'POST':
+            newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+            db.session.add(newItem)
+            db.session.commit()
+            flash('New Menu %s Item Successfully Created' % (newItem.name))
+            return redirect(url_for('main.showMenu', restaurant_id = restaurant_id))
+        else:
+            return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+    else: 
+       return redirect(url_for('main.showRestaurants'))
 
 #Edit a menu item
 @main.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
@@ -86,32 +101,37 @@ def editMenuItem(restaurant_id, menu_id):
 
     editedItem = db.session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
-    if request.method == 'POST':
-        if request.form['name']:
-            editedItem.name = request.form['name']
-        if request.form['description']:
-            editedItem.description = request.form['description']
-        if request.form['price']:
-            editedItem.price = request.form['price']
-        if request.form['course']:
-            editedItem.course = request.form['course']
-        db.session.add(editedItem)
-        db.session.commit() 
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('main.showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
-
+    if current_user.name == 'Admin' or current_user.id == restaurant.user_id:
+        if request.method == 'POST':
+            if request.form['name']:
+                editedItem.name = request.form['name']
+            if request.form['description']:
+                editedItem.description = request.form['description']
+            if request.form['price']:
+                editedItem.price = request.form['price']
+            if request.form['course']:
+                editedItem.course = request.form['course']
+            db.session.add(editedItem)
+            db.session.commit() 
+            flash('Menu Item Successfully Edited')
+            return redirect(url_for('main.showMenu', restaurant_id = restaurant_id))
+        else:
+            return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
+    else: 
+       return redirect(url_for('main.showRestaurants'))
 
 #Delete a menu item
 @main.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
     restaurant = db.session.query(Restaurant).filter_by(id = restaurant_id).one()
-    itemToDelete = db.session.query(MenuItem).filter_by(id = menu_id).one() 
-    if request.method == 'POST':
-        db.session.delete(itemToDelete)
-        db.session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('main.showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
+    if current_user.name == 'Admin' or current_user.id == restaurant.user_id:
+        itemToDelete = db.session.query(MenuItem).filter_by(id = menu_id).one() 
+        if request.method == 'POST':
+            db.session.delete(itemToDelete)
+            db.session.commit()
+            flash('Menu Item Successfully Deleted')
+            return redirect(url_for('main.showMenu', restaurant_id = restaurant_id))
+        else:
+            return render_template('deleteMenuItem.html', item = itemToDelete)
+    else: 
+       return redirect(url_for('main.showRestaurants'))
